+++ composeApp/src/commonMain/kotlin/com/poli/health/aquamate/onboarding/splash/di/SplashModule.kt
@@ -1,0 +1,10 @@
+package com.poli.health.aquamate.onboarding.splash.di
+
+import org.koin.core.module.Module
+import org.koin.dsl.module
+
+val splashModule: Module = module {
+    includes(
+        splashPresentationModule
+    )
+}