+++ composeApp/src/commonMain/kotlin/com/poli/health/aquamate/onboarding/splash/di/SplashPresentationModule.kt
@@ -1,0 +1,15 @@
+package com.poli.health.aquamate.onboarding.splash.di
+
+import com.poli.health.aquamate.onboarding.splash.presentation.viewmodel.SplashViewModel
+import org.koin.core.module.dsl.viewModel
+import org.koin.dsl.module
+
+val splashPresentationModule = module {
+    viewModel {
+        SplashViewModel(
+            isUserLoggedInUseCase = get(),
+            getCurrentUserIdUseCase = get(),
+            hasUserProfileUseCase = get()
+        )
+    }
+}